file-browser/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.0.0
pillow>=10.0.0
python-docx>=0.8.11
mammoth>=1.6.0
PyPDF2>=3.0.0
python-magic-bin>=0.4.14;sys_platform=="win32"
python-magic>=0.4.27;sys_platform!="win32"
